third_party/blink/web_tests/editing/input/edit-context.html

<html>
<head>
<title>This tests the new EditContext APIs while in composition</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-actions.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
</head>
<body>
<div id="test"></div>
<div id="test2"></div>
<script>

test(function() {
  const iframe = document.createElement("iframe");
  document.body.appendChild(iframe);
  const childDocument = iframe.contentDocument;
  const childDiv = childDocument.createElement('div');
  iframe.remove();
  childDiv.editContext = new EditContext();
}, 'Setting .editContext on a stray div should not crash.');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.createElement("div");
  document.body.appendChild(test);
  test.tabIndex = "0";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase() + buffer.substr(e.updateRangeEnd);
  });
  test.focus();
  test.editContext = editContext;
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  test.remove();
}, 'Testing attaching EditContext AFTER the element is focused.');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase() + buffer.substr(e.updateRangeEnd);
  });
  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.setComposition("bar");
  assert_equals(test.innerHTML, "BAR");
}, 'Testing EditContext TextUpdate');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "ABCD";
  editContext.updateText(0, 4, "abcd");
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase() + buffer.substr(e.updateRangeEnd);
  });

  let compositionStartCount = 0;
  let compositionEndCount = 0;
  editContext.addEventListener("compositionstart", e => {
    compositionStartCount++;
  });
  editContext.addEventListener("compositionend", e => {
    compositionEndCount++;
  });

  test.editContext = editContext;
  test.focus();
  editContext.updateSelection(2, 2);
  textInputController.setComposition("foo");
  assert_equals(editContext.text, "abfoocd");
  assert_equals(test.innerHTML, "ABFOOCD");
  assert_equals(compositionStartCount, 1, "Should get compositionstart after setting non-empty composition");
  assert_equals(compositionEndCount, 0, "Should not get compositionend until terminating composition");

  textInputController.setComposition("");
  assert_equals(editContext.text, "abcd");
  assert_equals(test.innerHTML, "ABCD");
  assert_equals(compositionStartCount, 1, "Shouldn't get another compositionstart when setting empty composition");
  assert_equals(compositionEndCount, 1, "Should get compositionend when setting empty composition");

  textInputController.setComposition("bar");
  assert_equals(editContext.text, "abbarcd");
  assert_equals(test.innerHTML, "ABBARCD");
  assert_equals(compositionStartCount, 2, "Should get another compositionstart after setting non-empty composition");
  assert_equals(compositionEndCount, 1, "Should not get another compositionend until terminating composition");
}, 'Testing EditContext TextUpdate with empty string');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  let textFormats = [];
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });
  editContext.addEventListener("textformatupdate", e => {
    //TODO: Currently Chromium only fires default styles
    textFormats = e.getTextFormats();
  });
  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  assert_equals(textFormats.length, 1);
  assert_equals(textFormats[0].rangeStart, 0);
  assert_equals(textFormats[0].rangeEnd, 3);
  assert_equals(textFormats[0].underlineStyle, "Solid");
  assert_equals(textFormats[0].underlineThickness, "Thin");
}, 'Testing EditContext TextFormatUpdate');

test(function() {
  const editContext1 = new EditContext();
  const editContext2 = new EditContext();
  assert_not_equals(editContext1, null);
  assert_not_equals(editContext2, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext1.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });
  editContext2.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toLowerCase()  + buffer.substr(e.updateRangeEnd);
  });
  test.editContext = editContext1;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.setComposition("bar");
  assert_equals(test.innerHTML, "BAR");
  test.innerHTML = "";
  test.editContext = editContext2;
  textInputController.setComposition("HELLO");
  assert_equals(test.innerHTML, "hello");
  textInputController.setComposition("WORLD");
  assert_equals(test.innerHTML, "world");
}, 'Testing Multiple EditContext TextUpdates');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  var compositionStartFired = 0;
  var compositionEndFired = 0;
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });
  editContext.addEventListener("compositionstart", e => {
    // Update the text in the div
    compositionStartFired++;
  });
  editContext.addEventListener("compositionend", e => {
    compositionEndFired++;
  });
  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  assert_equals(compositionStartFired, 1);
  textInputController.setComposition("bar");
  assert_equals(test.innerHTML, "BAR");
  assert_equals(compositionStartFired, 1);
  textInputController.insertText("bar");
  assert_equals(test.innerHTML, "BAR");
  assert_equals(compositionEndFired, 1);
}, 'Testing EditContext Composition Event');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("f");
  assert_equals(test.innerHTML, "F");
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.insertText("foobar");
  assert_equals(test.innerHTML, "FOOBAR");
}, 'Testing EditContext Text updates');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("f");
  assert_equals(test.innerHTML, "F");
  textInputController.setComposition("");
  assert_equals(test.innerHTML, "");
  textInputController.insertText("foobar");
  assert_equals(test.innerHTML, "FOOBAR");
}, 'Testing EditContext Text updates with empty text');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  editContext.updateText(0, 1, "h");
  assert_equals(editContext.text, "hoo");
  editContext.updateText(0, 3, "bar");
  assert_equals(editContext.text, "bar");
}, 'Testing EditContext TextChange');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  test.editContext = editContext;
  test.focus();
  assert_equals(editContext.text, "");
  editContext.updateText(0, 0, "foo");
  assert_equals(editContext.text, "foo");
  // replace "oo" with "bar"
  textInputController.setCompositionWithReplacementRange("bar", 1, 3);
  assert_equals(editContext.text, "fbar");
}, 'The new text should apply to the replacement range');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.insertText("bar");
  assert_equals(editContext.text, "bar");
  assert_equals(editContext.selectionStart, 3);
  assert_equals(editContext.selectionEnd, 3);
  editContext.updateSelection(0, 0);
  assert_equals(editContext.selectionStart, 0);
  assert_equals(editContext.selectionEnd, 0);
  textInputController.setComposition("foo");
  assert_equals(editContext.text, "foobar");
  textInputController.insertText("foo");
  assert_equals(editContext.selectionStart, 3);
  assert_equals(editContext.selectionEnd, 3);
}, 'Testing EditContext Selection Change');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  editContext.updateText(0, 1, "h");
  assert_equals(editContext.text, "hoo");
  editContext.updateText(0, 3, "bar");
  assert_equals(editContext.text, "bar");
  textInputController.insertText("bar");
  assert_equals(editContext.text, "bar");
  editContext.updateSelection(0, 0);
  assert_equals(editContext.selectionStart, 0);
  assert_equals(editContext.selectionEnd, 0);
  textInputController.setComposition("foo");
  assert_equals(editContext.text, "foobar");
  textInputController.insertText("foo");
  assert_equals(editContext.text, "foobar");
}, 'Testing EditContext Text And Selection Change');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.insertText("foo");
  textInputController.setMarkedTextFromExistingText(0, 3);
  textInputController.insertText("bar");
  assert_equals(editContext.text, "bar");
  textInputController.setMarkedTextFromExistingText(2, 3);
  textInputController.insertText("t");
  assert_equals(editContext.text, "bat");
}, 'Testing EditContext SetCompositionFromExistingText');

promise_test(async function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.insertText("foo");
  assert_equals(editContext.text, "foo", "EditContext's text should have been updated");
  assert_equals(editContext.selectionStart, 3, "EditContext's selectionStart should have been placed after inserted text");
  assert_equals(editContext.selectionEnd, 3, "EditContext's selectionEnd should have been placed after inserted text");

  textInputController.setMarkedTextFromExistingText(0, 3);
  assert_equals(editContext.text, "foo", "SetCompositionFromExistingText should not change EditContext's text");
  assert_equals(editContext.selectionStart, 3, "SetCompositionFromExistingText should not change EditContext's text");
  assert_equals(editContext.selectionEnd, 3, "SetCompositionFromExistingText should not change EditContext's text");

  await test_driver.send_keys(test, 'a');
  assert_equals(editContext.text, "fooa", "Normal typing should be inserted at selection position, not composition position");
  assert_equals(test.innerHTML, "FOOA",  "Normal typing should be inserted at selection position, not composition position");
}, 'Testing EditContext SetCompositionFromExistingText followed by non-composition typing');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.insertText("foo");
  assert_equals(editContext.selectionStart, 3);
  assert_equals(editContext.selectionEnd, 3);
  textInputController.extendSelectionAndDelete(3, 0);
  assert_equals(editContext.selectionStart, 0);
  assert_equals(editContext.selectionEnd, 0);
  textInputController.insertText("bar");
  assert_equals(editContext.text, "bar");
  assert_equals(editContext.selectionStart, 3);
  assert_equals(editContext.selectionEnd, 3);
  textInputController.extendSelectionAndDelete(1, 0);
  assert_equals(editContext.selectionStart, 2);
  assert_equals(editContext.selectionEnd, 2);
  assert_equals(editContext.text, "ba");
  textInputController.insertText("t");
  assert_equals(editContext.text, "bat");
}, 'Testing EditContext ExtendSelectionAndDelete');

test(function() {
  const editContext = new EditContext();
  assert_not_equals(editContext, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  textInputController.unmarkText();
  assert_equals(editContext.text, "foo");
  textInputController.setComposition("bar");
  assert_equals(editContext.text, "foobar");
}, 'Testing EditContext FinishComposingText');

test(function() {
  const editContext1 = new EditContext();
  assert_not_equals(editContext1, null);
  const editContext2 = new EditContext();
  assert_not_equals(editContext2, null);
  const test = document.getElementById('test');
  test.innerHTML = "";
  // Add EditContext event listeners
  editContext1.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toUpperCase()  + buffer.substr(e.updateRangeEnd);
  });

  editContext2.addEventListener("textupdate", e => {
    // Update the text in the div
    const buffer = test.innerText;
    test.innerHTML = buffer.substr(0, e.updateRangeStart) + e.text.toLowerCase()  + buffer.substr(e.updateRangeEnd);
  });

  test.editContext = editContext1;
  test.focus();
  textInputController.setComposition("foo");
  assert_equals(test.innerHTML, "FOO");
  test.blur();
  textInputController.setComposition("foo2");
  assert_equals(test.innerHTML, "FOO");

  const test2 = document.getElementById('test2');
  test2.editContext = editContext2;
  test2.focus();
  textInputController.setComposition("BAR");
  assert_equals(editContext2.text, "BAR");
  assert_equals(test.innerHTML, "barFOO");
  assert_equals(editContext1.text, "foo");
}, 'Testing EditContext blur');

test(function() {
  // SetComposition should not crash when event handler removes document
  const child = document.createElement("iframe");
  document.body.appendChild(child);
  const childDocument = child.contentDocument;
  const editable = childDocument.createElement('div');
  editable.contentEditable = true;
  childDocument.body.appendChild(editable);
  editable.addEventListener("focusin", e => {
    const childEditContext = new EditContext();
    editable.editContext = childEditContext;
    childEditContext.addEventListener("textupdate", e => {
      child.remove();
    });
    childEditContext.addEventListener("textformatupdate", e => {
    });
  });
  editable.focus();
  child.contentWindow.focus();
  child.contentWindow.textInputController.setComposition("bar");
}, 'Testing EditContext Iframe Document Delete');

test(function() {
  const editContext1 = new EditContext();
  editContext1.addEventListener("textupdate", e => {
  });
  const test = document.getElementById('test');
  test.editContext = editContext1;
  test.focus();
  gc()
  textInputController.setComposition("bar");

}, 'Testing EditContext GC');
</script>
</body>
</html>
